ultimo comit

+++ demo/src/main/java/com/demo/repository/ArticuloRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.model.Articulo;
 
-public interface ArticuloRepository extends JpaRepository<Articulo, Long> {
+public interface ArticuloRepository extends JpaRepository<Articulo, Integer> {
 
 }
